GD plugin, glcomp, smyrna modernization

See merge request graphviz/graphviz!4515

+++ lib/glcomp/glcompfont.h
@@ -27,7 +27,7 @@ glCompFont glNewFont(glCompSet *s, const char *text, glCompColor *c,
                      char *fontdesc, int fs, bool is2D);
 glCompFont glNewFontFromParent(glCompObj *o, const char *text);
     void glDeleteFont(glCompFont * f);
-void glCompDrawText(glCompFont f, float x, float y);
+void glCompDrawText(glCompFont f, double x, double y);
 void glCompRenderText(glCompFont f, glCompObj *parentObj);
 void glCompDrawText3D(glCompFont f, float x, float y, double z, float w,
                       float h);
